move utils to separate files
